- opcja usuwanie logów z historii - wyświetlanie historii logów dla danego użytkownika - zapisywanie logow wyszukiwania do bazy danych - komunikaty błednego logowania w messages.properties - hashowanie hasła, zmiana logowania i rejestracji na LoginForm itp. - wykres google charts - nie działa pobieranie danych z jsona do wykresu - działa przed dodaniem wykresu 5 dni - logowanie użytkowników - prognoza na 5 dni (tabela)

+++ src/main/java/pl/lukabrasi/pogodaonline/weather/controllers/HistoryController.java
@@ -25,44 +25,26 @@ public class HistoryController {
 
     @GetMapping("/history")
     public String histoty(Model model) {
+        if (!userSession.isLogin()) {
+            return "redirect:/login";
+        }
         model.addAttribute("logs", weatherLogService.getWeatherLogFromUser());
         return "history";
     }
 
 
-/*    @DeleteMapping(value = "/history/{id}")
-    public String delete(@PathVariable(value = "id") int id) {
-        weatherLogService.deleteWeatherLogEntityById(id);
-        return "redirect:/history";
-    }*/
 
     @RequestMapping(value = "/history/{id}", method = RequestMethod.GET)
     public String delete(@PathVariable(value = "id") int id) {
-
+        if (!userSession.isLogin()) {
+            return "redirect:/login";
+        }
         weatherLogService.deleteWeatherLogEntityById(id);
-        weatherLogService.getWeatherLogFromUser();
+
 
         return "redirect:/history";
     }
 
-
-/*  @DeleteMapping("/delete/{id}")
-    public ResponseEntity<Void> deleteWeatherLogEntity(@PathVariable("id") Integer id) {
-        weatherLogService.deleteWeatherLogEntity(id);
-        return new ResponseEntity<Void>(HttpStatus.NO_CONTENT);
-
-    }*/
-
-
-
-/*    @GetMapping("/delete/{id}")
-    public String deleteUser(@PathVariable("id") long id, Model model) {
-        User user = userRepository.findById(id)
-                .orElseThrow(() -> new IllegalArgumentException("Invalid user Id:" + id));
-        userRepository.delete(user);
-        model.addAttribute("users", userRepository.findAll());
-        return "index";
-    }*/
 
 
 }
